[#617] Move fake navigator, resolve more comments
